Moved reference_test_utilities to test_utilities
