+++ plugins/miopen_legacy_plugin/tests/TestMiopenBatchnormFwdInferenceOperations.cpp
@@ -18,7 +18,7 @@
 #include "HipdnnEnginePluginHandle.hpp"
 #include "common/TestOperationsCommon.hpp"
 
-using namespace hipdnn_sdk::reference_test_utilities;
+using namespace hipdnn_sdk::test_utilities;
 using namespace test_operations_common;
 
 class BatchnormFwdInferExecuteGraphTest : public ::testing::TestWithParam<Batchnorm2dTestCase>
