+++ samples/batchnorm/BnTraining.cpp
@@ -131,8 +131,8 @@ void SampleRunner::operator()(const TensorLayout& layout)
     {
         std::cout << "Running CPU reference validation...\n";
 
-        auto refImpl = hipdnn_sdk::reference_test_utilities::
-            CpuFpReferenceImplementation<InputType, IntermediateType>();
+        auto refImpl = hipdnn_sdk::test_utilities::CpuFpReferenceImplementation<InputType,
+                                                                                IntermediateType>();
 
         Tensor<InputType> yRefTensor(y->get_dim(), layout);
         Tensor<IntermediateType> nextMeanRefTensor(nextRunningMean->get_dim());
@@ -157,11 +157,11 @@ void SampleRunner::operator()(const TensorLayout& layout)
         // auto epsilon = get_epsilon<InputType>();
         //
         // auto y_validator
-        //     = hipdnn_sdk::reference_test_utilities::CpuFpReferenceValidation<InputType>(
+        //     = hipdnn_sdk::test_utilities::CpuFpReferenceValidation<InputType>(
         //         static_cast<InputType>(epsilon), static_cast<InputType>(epsilon));
         //
         // auto stats_validator
-        //     = hipdnn_sdk::reference_test_utilities::CpuFpReferenceValidation<IntermediateType>(
+        //     = hipdnn_sdk::test_utilities::CpuFpReferenceValidation<IntermediateType>(
         //         static_cast<IntermediateType>(epsilon), static_cast<IntermediateType>(epsilon));
         // bool y_valid = y_validator.allClose(y_ref_tensor.memory(), y_tensor.memory());
         // bool next_mean_valid = stats_validator.allClose(next_mean_ref_tensor.memory(),
