+++ sdk/include/hipdnn_sdk/test_utilities/CpuFpReferenceImplementation.hpp
@@ -15,7 +15,7 @@
 
 namespace hipdnn_sdk
 {
-namespace reference_test_utilities
+namespace test_utilities
 {
 
 using namespace hipdnn_sdk::utilities;
@@ -187,5 +187,5 @@ private:
     }
 };
 
-} // namespace reference_test_utilities
+} // namespace test_utilities
 } // namespace hipdnn_sdk
