+++ sdk/include/hipdnn_sdk/test_utilities/CpuFpReferenceMiopenRmsValidation.hpp
@@ -14,7 +14,7 @@
 
 namespace hipdnn_sdk
 {
-namespace reference_test_utilities
+namespace test_utilities
 {
 
 using namespace hipdnn_sdk::utilities;
@@ -91,5 +91,5 @@ private:
     double _relativeTolerance;
 };
 
-} // namespace reference_test_utilities
+} // namespace test_utilities
 } // namespace hipdnn_sdk
