+++ sdk/include/hipdnn_sdk/test_utilities/CpuFpReferenceValidation.hpp
@@ -14,7 +14,7 @@
 
 namespace hipdnn_sdk
 {
-namespace reference_test_utilities
+namespace test_utilities
 {
 
 using namespace hipdnn_sdk::utilities;
@@ -81,5 +81,5 @@ private:
     T _relativeTolerance;
 };
 
-} // namespace reference_test_utilities
+} // namespace test_utilities
 } // namespace hipdnn_sdk
