+++ sdk/include/hipdnn_sdk/test_utilities/ReferenceImplementationInterface.hpp
@@ -13,7 +13,7 @@
 
 namespace hipdnn_sdk
 {
-namespace reference_test_utilities
+namespace test_utilities
 {
 
 using namespace hipdnn_sdk::data_objects;
@@ -48,7 +48,7 @@ public:
         = 0;
 };
 
-} // namespace reference_test_utilities
+} // namespace test_utilities
 } // namespace hipdnn_sdk
 
 // NOLINTEND(portability-template-virtual-member-function)
