+++ sdk/tests/test_utilities/TestCpuFpReferenceImplementation.cpp
@@ -10,7 +10,7 @@
 #include <hipdnn_sdk/utilities/HipBfloat16Utils.hpp>
 #include <hipdnn_sdk/utilities/Tensor.hpp>
 
-using namespace hipdnn_sdk::reference_test_utilities;
+using namespace hipdnn_sdk::test_utilities;
 using namespace hipdnn_sdk::data_objects;
 using namespace hipdnn_sdk::utilities;
